+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -364,6 +364,8 @@ public class SampleSQLBuilder {
 			"journalDataFactory", new JournalDataFactory()
 		).put(
 			"messageBoardDataFactory", new MessageBoardDataFactory()
+		).put(
+			"wikiDataFactory", new WikiDataFactory()
 		).build();
 	}
