+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/WikiDataFactory.java
@@ -1,0 +1,130 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.sample.sql.builder;
+
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.wiki.model.WikiNodeModel;
+import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.model.WikiPageConstants;
+import com.liferay.wiki.model.WikiPageModel;
+import com.liferay.wiki.model.WikiPageResourceModel;
+import com.liferay.wiki.model.impl.WikiNodeModelImpl;
+import com.liferay.wiki.model.impl.WikiPageModelImpl;
+import com.liferay.wiki.model.impl.WikiPageResourceModelImpl;
+
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
+/**
+ * @author Lily Chi
+ */
+public class WikiDataFactory extends BaseDataFactory {
+
+	public int getMaxWikiPageCommentCount() {
+		return PropsValues.MAX_WIKI_PAGE_COMMENT_COUNT;
+	}
+
+	public long getWikiPageClassNameId() {
+		return getClassNameId(WikiPage.class);
+	}
+
+	public List<WikiNodeModel> newWikiNodeModels(long groupId) {
+		List<WikiNodeModel> wikiNodeModels = new ArrayList<>(
+			PropsValues.MAX_WIKI_NODE_COUNT);
+
+		for (int i = 1; i <= PropsValues.MAX_WIKI_NODE_COUNT; i++) {
+			wikiNodeModels.add(newWikiNodeModel(groupId, i));
+		}
+
+		return wikiNodeModels;
+	}
+
+	public List<WikiPageModel> newWikiPageModels(WikiNodeModel wikiNodeModel) {
+		List<WikiPageModel> wikiPageModels = new ArrayList<>(
+			PropsValues.MAX_WIKI_PAGE_COUNT);
+
+		for (int i = 1; i <= PropsValues.MAX_WIKI_PAGE_COUNT; i++) {
+			wikiPageModels.add(newWikiPageModel(wikiNodeModel, i));
+		}
+
+		return wikiPageModels;
+	}
+
+	public WikiPageResourceModel newWikiPageResourceModel(
+		WikiPageModel wikiPageModel) {
+
+		WikiPageResourceModel wikiPageResourceModel =
+			new WikiPageResourceModelImpl();
+
+		wikiPageResourceModel.setUuid(SequentialUUID.generate());
+		wikiPageResourceModel.setResourcePrimKey(
+			wikiPageModel.getResourcePrimKey());
+		wikiPageResourceModel.setNodeId(wikiPageModel.getNodeId());
+		wikiPageResourceModel.setTitle(wikiPageModel.getTitle());
+
+		return wikiPageResourceModel;
+	}
+
+	protected WikiNodeModel newWikiNodeModel(long groupId, int index) {
+		WikiNodeModel wikiNodeModel = new WikiNodeModelImpl();
+
+		wikiNodeModel.setUuid(SequentialUUID.generate());
+		wikiNodeModel.setNodeId(counter.get());
+		wikiNodeModel.setGroupId(groupId);
+		wikiNodeModel.setCompanyId(COMPANY_ID);
+		wikiNodeModel.setUserId(SAMPLE_USER_ID);
+		wikiNodeModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
+		wikiNodeModel.setCreateDate(new Date());
+		wikiNodeModel.setModifiedDate(new Date());
+		wikiNodeModel.setName(
+			DataFactoryConstants.WIKI_NODE_NAME_PREFIX + index);
+		wikiNodeModel.setLastPostDate(new Date());
+		wikiNodeModel.setLastPublishDate(new Date());
+		wikiNodeModel.setStatusDate(new Date());
+
+		return wikiNodeModel;
+	}
+
+	protected WikiPageModel newWikiPageModel(
+		WikiNodeModel wikiNodeModel, int index) {
+
+		WikiPageModel wikiPageModel = new WikiPageModelImpl();
+
+		wikiPageModel.setUuid(SequentialUUID.generate());
+		wikiPageModel.setPageId(counter.get());
+		wikiPageModel.setResourcePrimKey(counter.get());
+		wikiPageModel.setGroupId(wikiNodeModel.getGroupId());
+		wikiPageModel.setCompanyId(COMPANY_ID);
+		wikiPageModel.setUserId(SAMPLE_USER_ID);
+		wikiPageModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
+		wikiPageModel.setCreateDate(new Date());
+		wikiPageModel.setModifiedDate(new Date());
+		wikiPageModel.setNodeId(wikiNodeModel.getNodeId());
+		wikiPageModel.setTitle(
+			DataFactoryConstants.WIKI_PAGE_TITLE_PREFIX + index);
+		wikiPageModel.setVersion(WikiPageConstants.VERSION_DEFAULT);
+		wikiPageModel.setContent(
+			StringBundler.concat(
+				DataFactoryConstants.WIKI_PAGE_CONTENT_PREFIX,
+				String.valueOf(index), " of ", wikiNodeModel.getName(), "."));
+		wikiPageModel.setFormat(DataFactoryConstants.WIKI_PAGE_FORMAT);
+		wikiPageModel.setHead(true);
+		wikiPageModel.setLastPublishDate(new Date());
+
+		return wikiPageModel;
+	}
+
+}
